pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-foresight"
version = "0.1.0"
description = "Multi-country OSINT foresight stack"
authors = [{name = "Your Team"}]
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.2",
  "PyYAML>=6.0.1",
  "requests>=2.32.3",
  "python-dotenv>=1.0.1",
  "tqdm>=4.66.4",
  "rich>=13.7.1",
]

[tool.setuptools.packages.find]
where = ["src"]